[refactor] 최적의 회의 시간대 반환 로직 분리
