+++ src/main/java/com/asap/server/common/utils/BestMeetingUtil.java
@@ -1,11 +1,10 @@
 package com.asap.server.common.utils;
 
 import com.asap.server.common.utils.strategy.FindOptimalMeetingTimeCasesStrategy;
+import com.asap.server.common.utils.strategy.FindOptimalMeetingTimeStrategy;
 import com.asap.server.domain.enums.Duration;
-import com.asap.server.domain.enums.TimeSlot;
 import com.asap.server.service.vo.BestMeetingTimeVo;
 import com.asap.server.service.vo.PossibleTimeCaseVo;
-import com.asap.server.service.vo.TimeBlockVo;
 import com.asap.server.service.vo.TimeBlocksByDateVo;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +20,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class BestMeetingUtil {
     private final FindOptimalMeetingTimeCasesStrategy findOptimalMeetingTimeCasesStrategy;
+    private final FindOptimalMeetingTimeStrategy findOptimalMeetingTimeStrategy;
 
     public List<BestMeetingTimeVo> getBestMeetingTime(
             final List<TimeBlocksByDateVo> timeBlocksByDates,
@@ -32,9 +32,13 @@ public class BestMeetingUtil {
 
         for (PossibleTimeCaseVo timeCase : timeCases) {
             timeBlocksByDates.forEach(timeBlocksByDate -> {
-                bestMeetingTimes.addAll(
-                        searchBestMeetingTime(timeBlocksByDate, timeCase.getDuration().getNeedBlock(), timeCase.getMemberCnt())
+                List<BestMeetingTimeVo> bestMeetingTimeVos = findOptimalMeetingTimeStrategy.findOptimalMeetingTime(
+                        timeBlocksByDate,
+                        timeCase.getDuration().getNeedBlock(),
+                        timeCase.getMemberCnt()
                 );
+                
+                bestMeetingTimes.addAll(bestMeetingTimeVos);
             });
             if (bestMeetingTimes.size() > 2) return bestMeetingTimes
                     .stream()
@@ -48,53 +52,4 @@ public class BestMeetingUtil {
         }
         return bestMeetingTimes;
     }
-
-    public List<BestMeetingTimeVo> searchBestMeetingTime(final TimeBlocksByDateVo timeBlocksByDate, final int needTimeBlockCount, final int userCount) {
-        List<TimeBlockVo> sortedTimeBlocks = filterByUserCountAndSortByTime(timeBlocksByDate.getTimeBlocks(), userCount);
-
-        List<BestMeetingTimeVo> bestMeetingTimes = new ArrayList<>();
-        int endIndex = sortedTimeBlocks.size() - needTimeBlockCount + 1;
-        for (int timeBlockIdx = 0; timeBlockIdx < endIndex; timeBlockIdx++) {
-            if (!isBestMeetingTime(sortedTimeBlocks, timeBlockIdx, needTimeBlockCount)) continue;
-
-            int sumWeight = sortedTimeBlocks
-                    .subList(timeBlockIdx, timeBlockIdx + needTimeBlockCount)
-                    .stream()
-                    .map(TimeBlockVo::getWeight)
-                    .reduce(0, Integer::sum);
-
-            TimeSlot startTime = sortedTimeBlocks.get(timeBlockIdx).getTimeSlot();
-            BestMeetingTimeVo bestMeetingTime = new BestMeetingTimeVo(
-                    timeBlocksByDate.getDate(),
-                    startTime,
-                    TimeSlot.getTimeSlot(startTime.ordinal() + needTimeBlockCount),
-                    sortedTimeBlocks.get(timeBlockIdx).getUsers(),
-                    sumWeight
-            );
-            bestMeetingTimes.add(bestMeetingTime);
-        }
-
-        return bestMeetingTimes;
-    }
-
-    private List<TimeBlockVo> filterByUserCountAndSortByTime(final List<TimeBlockVo> timeBlocks, final int userCount) {
-        return timeBlocks.stream()
-                .filter(timeBlockVo -> timeBlockVo.getUsers().size() == userCount)
-                .sorted(TimeBlockVo::compareTo)
-                .collect(Collectors.toList());
-    }
-
-    private boolean isBestMeetingTime(final List<TimeBlockVo> timeBlocks, final int timeBlockIdx, final int needTimeBlockCount) {
-        boolean isBestMeetingTime = true;
-        TimeSlot nextTime = timeBlocks.get(timeBlockIdx).getTimeSlot();
-        for (int i = timeBlockIdx + 1; i < timeBlockIdx + needTimeBlockCount; i++) {
-            if (nextTime.ordinal() + 1 != timeBlocks.get(i).getTimeSlot().ordinal()) {
-                isBestMeetingTime = false;
-                break;
-            }
-            nextTime = timeBlocks.get(i).getTimeSlot();
-        }
-        return isBestMeetingTime;
-    }
-
 }
