+++ src/test/java/com/asap/server/common/utils/GetBestMeetingTimeTest.java
@@ -1,6 +1,7 @@
 package com.asap.server.common.utils;
 
 import com.asap.server.common.utils.strategy.impl.FindOptimalMeetingTimeCasesStrategyImpl;
+import com.asap.server.common.utils.strategy.impl.FindOptimalMeetingTimeStrategyImpl;
 import com.asap.server.domain.enums.Duration;
 import com.asap.server.service.vo.BestMeetingTimeVo;
 import com.asap.server.service.vo.TimeBlockVo;
@@ -28,7 +29,10 @@ public class GetBestMeetingTimeTest {
 
     @BeforeEach
     public void setUp() {
-        bestMeetingUtil = new BestMeetingUtil(new FindOptimalMeetingTimeCasesStrategyImpl());
+        bestMeetingUtil = new BestMeetingUtil(
+                new FindOptimalMeetingTimeCasesStrategyImpl(),
+                new FindOptimalMeetingTimeStrategyImpl()
+        );
     }
 
     @Test
